Fix and migrate `tests/unit/cloud/clouds/test_gce.py` to PyTest
